requirements/base.txt
django==1.10.3
django-allauth==0.29.0
djangorestframework==3.5.1
djet==0.1.0
factory-boy==2.7.0
faker==0.7.3
Pillow==3.4.2
